// projects/myfavors/src/app/enums/myfavors.enum.ts
export enum MYFAVORS_EVENT_TYPES {
  FAVOR_NEW_UPDATE = 'FAVOR_NEW_UPDATE',
  FAVOR_NEW_MESSAGE = 'FAVOR_NEW_MESSAGE',
  FAVOR_STARTED = 'FAVOR_STARTED',
  FAVOR_FULFILLED = 'FAVOR_FULFILLED',
  FAVOR_CANCELED = 'FAVOR_CANCELED',
  FAVOR_HELPER_REQUESTED = 'FAVOR_HELPER_REQUESTED',
  FAVOR_HELPER_HELPED = 'FAVOR_HELPER_HELPED',
  FAVOR_HELPER_UNHELPED = 'FAVOR_HELPER_UNHELPED',
  FAVOR_HELPER_PAID = 'FAVOR_HELPER_PAID',
  FAVOR_HELPER_UNREQUESTED = 'FAVOR_HELPER_UNREQUESTED',
  FAVOR_HELPER_ASSIGNED = 'FAVOR_HELPER_ASSIGNED',
  FAVOR_HELPER_UNASSIGNED = 'FAVOR_HELPER_UNASSIGNED',
  FAVOR_PAYOUT = 'FAVOR_PAYOUT',

  TO_FAVOR = 'TO_FAVOR',
}

export enum MYFAVORS_NOTIFICATION_TARGET_TYPES {
  FAVOR = 'FAVOR',
  FAVOR_UPDATE = 'FAVOR_UPDATE',
  FAVOR_MESSAGE = 'FAVOR_MESSAGE',
}